kafka2elasticsearch/requirements.txt
kafka-python==1.0.2
ipython
elasticsearch_async
